requirements.txt
pydantic==2.7.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
boto3==1.34.103
SQLAlchemy==2.0.30
uvicorn==0.29.0
fastapi==0.111.0
python-multipart==0.0.9
fastapi-crudrouter==0.8.6
psycopg2-binary==2.9.9
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.31.0
websockets==12.0
fastapi-mail==1.4.1
geoip2==4.8.0
pyotp==2.9.0
clamd==1.0.2
azure-storage-blob==12.19.1